requirements.txt
accelerate==0.34.2
datasets==2.21.0
matplotlib==3.9.2
numpy==1.26.4
pandas==2.2.3
PyYAML==6.0.2
rank-bm25==0.2.2
scikit-learn==1.5.2
torch==2.4.1
tqdm==4.66.5
transformers==4.45.2